refactor: AppContainer를 삭제하고 ShoppingApplication에서 의존성을 관리하도록 변경
