+++ app/src/main/java/woowacourse/shopping/feature/cart/CartViewModel.kt
@@ -1,6 +1,7 @@
 package woowacourse.shopping.feature.cart
 
 import androidx.lifecycle.ViewModel
+import androidx.lifecycle.ViewModelProvider.AndroidViewModelFactory.Companion.APPLICATION_KEY
 import androidx.lifecycle.viewModelScope
 import androidx.lifecycle.viewmodel.initializer
 import androidx.lifecycle.viewmodel.viewModelFactory
@@ -12,8 +13,7 @@ import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.receiveAsFlow
 import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
-import woowacourse.shopping.AppContainer
-import woowacourse.shopping.AppContainer.cartRepository
+import woowacourse.shopping.ShoppingApplication
 import woowacourse.shopping.data.repository.cart.CartRepository
 import woowacourse.shopping.domain.Cart
 import woowacourse.shopping.domain.CartContent
@@ -173,7 +173,8 @@ class CartViewModel(
     companion object {
         val Factory = viewModelFactory {
             initializer {
-                CartViewModel(5, AppContainer.cartRepository)
+                val app = this[APPLICATION_KEY] as ShoppingApplication
+                CartViewModel(5, app.cartRepository)
             }
         }
     }
